removed fetcher component from block service, remove Fetcher interface - no longer needed
