+++ node/bridge_test.go
@@ -99,7 +99,6 @@ func TestBridge_P2P_Streams(t *testing.T) {
 
 func TestBridge_WithRemoteCore(t *testing.T) {
 	// TODO(@Wondertan): Fix core
-	t.Skip("Skip until we fix core")
 
 	store := MockStore(t, DefaultConfig(Bridge))
 	remoteCore, protocol, ip := core.StartRemoteCore()
