+++ service/header/core_exchange.go
@@ -4,19 +4,16 @@ import (
 	"bytes"
 	"context"
 	"fmt"
-	"math"
 
 	format "github.com/ipfs/go-ipld-format"
+
 	tmbytes "github.com/tendermint/tendermint/libs/bytes"
 	"github.com/tendermint/tendermint/pkg/da"
-	"github.com/tendermint/tendermint/pkg/wrapper"
 	"github.com/tendermint/tendermint/types"
 
-	"github.com/celestiaorg/nmt"
-	"github.com/celestiaorg/rsmt2d"
-
 	"github.com/celestiaorg/celestia-node/core"
-	"github.com/celestiaorg/celestia-node/ipld"
+	"github.com/celestiaorg/celestia-node/utils"
+	"github.com/celestiaorg/rsmt2d"
 )
 
 type CoreExchange struct {
@@ -125,25 +122,7 @@ func (ce *CoreExchange) generateExtendedHeaderFromBlock(block *types.Block) (*Ex
 
 // extendBlockData erasure codes the given raw block's data and returns the
 // erasure coded block data upon success.
-// TODO @renaynay: this is duplicate code for purpose of DEVNET, delete post devnet
+// TODO @renaynay: this code is needed several places in the codebase and would be good to extract into a utility pkg
 func (ce *CoreExchange) extendBlockData(raw *types.Block) (*rsmt2d.ExtendedDataSquare, error) {
-	namespacedShares, _ := raw.Data.ComputeShares()
-	shares := namespacedShares.RawShares()
-
-	// create nmt adder wrapping batch adder
-	batchAdder := ipld.NewNmtNodeAdder(context.Background(), format.NewBatch(context.Background(), ce.shareStore))
-
-	// create the nmt wrapper to generate row and col commitments
-	squareSize := squareSize64(len(namespacedShares))
-	tree := wrapper.NewErasuredNamespacedMerkleTree(squareSize, nmt.NodeVisitor(batchAdder.Visit))
-
-	// compute extended square
-	return rsmt2d.ComputeExtendedDataSquare(shares, rsmt2d.NewRSGF8Codec(), tree.Constructor)
-}
-
-// squareSize64 computes the square size as a uint64 from
-// the given length of shares.
-// TODO @renaynay: this is duplicate code for purpose of DEVNET, delete post devnet
-func squareSize64(length int) uint64 {
-	return uint64(math.Sqrt(float64(length)))
+	return utils.ExtendBlock(raw, ce.shareStore)
 }
